Link to Chris Chester-Young's stream-constant.
Chris Chester-Young submitted a performance improvement to
stream-constant.  Phil Bewig, the author of SRFI 41, agreed that it is
an improvement.  However, since the SRFI has been finalized and this
isn't a fix for a trivial typo or similar error, it doesn't belong in
an errata release.  Add a link to Chris's message original instead.